series plot

+++ Magnetic_Field_Evaluation/series.py
@@ -4,10 +4,12 @@ import os
 import pickle
 
 from tool.metric import evaluate_single
+from tool.series_plot import plot_with_flares
 from tqdm import tqdm
 
 parser = argparse.ArgumentParser()
 
+parser.add_argument("noaanum", type=int, help="NOAA AR number")
 parser.add_argument(
     "--fname",
     type=str,
@@ -21,23 +23,35 @@ parser.add_argument(
 parser.add_argument(
     "--z_Mm", type=int, help="max height for z-direction [Mm]", default=40
 )
+parser.add_argument("--fig_dir", type=str, help="path to results.pickle", default=".")
 
 args = parser.parse_args()
 
-if os.path.isdir(args.fname):
-    files = sorted(glob.glob(os.path.join(args.fname, args.ext)))
-elif os.path.isfile(args.fname):
-    files = [args.fname]
-
-os.makedirs(args.result_dir, exist_ok=True)
-
-results = []
-
-for file in tqdm(files):
-    results.append(evaluate_single(file, args))
-
-results = {k: [r[k] for r in results] for k in results[0].keys()}
-
 result_pickle = os.path.join(args.result_dir, "series_result.pickle")
-with open(result_pickle, "wb") as f:
-    pickle.dump(results, f)
+
+if os.path.exists(result_pickle):
+    with open(result_pickle, "rb") as f:
+        series_results = pickle.load(f)
+
+    fig_path = os.path.join(args.fig_dir, f"NOAA_{args.noaanum}.png")
+    plot_with_flares(series_results, args.noaanum, fig_path)
+else:
+    if os.path.isdir(args.fname):
+        files = sorted(glob.glob(os.path.join(args.fname, args.ext)))
+    elif os.path.isfile(args.fname):
+        files = [args.fname]
+
+    os.makedirs(args.result_dir, exist_ok=True)
+
+    results = []
+
+    for file in tqdm(files):
+        results.append(evaluate_single(file, args))
+
+    results = {k: [r[k] for r in results] for k in results[0].keys()}
+
+    with open(result_pickle, "wb") as f:
+        pickle.dump(results, f)
+
+    fig_path = os.path.join(args.fig_dir, f"NOAA_{args.noaanum}.png")
+    plot_with_flares(results, args.noaanum, fig_path)
